% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iftp_distributions.R
\name{dist_arcsine}
\alias{dist_arcsine}
\title{Create an arcsine distribution specification}
\usage{
dist_arcsine()
}
\value{
An \code{iftp_dist} object with \code{cdf} and \code{quantile} functions.
}
\description{
The standard arcsine distribution on \eqn{[0, 1]} with CDF
\code{F(x) = (2/pi) * asin(sqrt(x))}.
}
\examples{
d <- dist_arcsine()
d$cdf(0.5)

}
\seealso{
Other distributions: 
\code{\link{dist_exponential}()},
\code{\link{dist_frechet}()},
\code{\link{dist_gamma}()},
\code{\link{dist_log_logistic}()},
\code{\link{dist_lomax}()},
\code{\link{dist_soft_killing_constant_barrier}()},
\code{\link{dist_uniform}()},
\code{\link{dist_weibull}()}
}
\concept{distributions}
