% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iftp_distributions.R
\name{dist_exponential}
\alias{dist_exponential}
\title{Create an exponential distribution specification}
\usage{
dist_exponential(rate = 1)
}
\arguments{
\item{rate}{Rate parameter (inverse mean). Defaults to 1.}
}
\value{
An \code{iftp_dist} object with \code{cdf} and \code{quantile} functions.
}
\description{
Create an exponential distribution specification
}
\examples{
d <- dist_exponential(rate = 2)
d$cdf(1)
d$qf(0.5)

}
\seealso{
Other distributions: 
\code{\link{dist_arcsine}()},
\code{\link{dist_frechet}()},
\code{\link{dist_gamma}()},
\code{\link{dist_log_logistic}()},
\code{\link{dist_lomax}()},
\code{\link{dist_soft_killing_constant_barrier}()},
\code{\link{dist_uniform}()},
\code{\link{dist_weibull}()}
}
\concept{distributions}
