% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iftp_distributions.R
\name{dist_frechet}
\alias{dist_frechet}
\title{Create a Frechet distribution specification}
\usage{
dist_frechet(shape = 1, scale = 1)
}
\arguments{
\item{shape}{Shape parameter (alpha). Defaults to 1.}

\item{scale}{Scale parameter (sigma). Defaults to 1.}
}
\value{
An \code{iftp_dist} object with \code{cdf} and \code{quantile} functions.
}
\description{
Type II extreme value distribution with CDF
\code{F(x) = exp(-(x/scale)^(-shape))} for \code{x > 0}.
}
\examples{
d <- dist_frechet(shape = 2)
d$cdf(1)

}
\seealso{
Other distributions: 
\code{\link{dist_arcsine}()},
\code{\link{dist_exponential}()},
\code{\link{dist_gamma}()},
\code{\link{dist_log_logistic}()},
\code{\link{dist_lomax}()},
\code{\link{dist_soft_killing_constant_barrier}()},
\code{\link{dist_uniform}()},
\code{\link{dist_weibull}()}
}
\concept{distributions}
