% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iftp_distributions.R
\name{dist_gamma}
\alias{dist_gamma}
\title{Create a Gamma distribution specification}
\usage{
dist_gamma(shape = 1, rate = 1)
}
\arguments{
\item{shape}{Shape parameter. Defaults to 1.}

\item{rate}{Rate parameter (inverse scale). Defaults to 1.}
}
\value{
An \code{iftp_dist} object with \code{cdf} and \code{quantile} functions.
}
\description{
Create a Gamma distribution specification
}
\examples{
d <- dist_gamma(shape = 2, rate = 1)
d$cdf(1)

}
\seealso{
Other distributions: 
\code{\link{dist_arcsine}()},
\code{\link{dist_exponential}()},
\code{\link{dist_frechet}()},
\code{\link{dist_log_logistic}()},
\code{\link{dist_lomax}()},
\code{\link{dist_soft_killing_constant_barrier}()},
\code{\link{dist_uniform}()},
\code{\link{dist_weibull}()}
}
\concept{distributions}
