% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iftp_distributions.R
\name{dist_log_logistic}
\alias{dist_log_logistic}
\title{Create a log-logistic distribution specification}
\usage{
dist_log_logistic(shape = 1, scale = 1)
}
\arguments{
\item{shape}{Shape parameter. Defaults to 1.}

\item{scale}{Scale parameter. Defaults to 1.}
}
\value{
An \code{iftp_dist} object with \code{cdf} and \code{quantile} functions.
}
\description{
Requires the \pkg{flexsurv} package at runtime.
}
\examples{
if (requireNamespace("flexsurv", quietly = TRUE)) {
  d <- dist_log_logistic(shape = 8)
  d$cdf(1)
}

}
\seealso{
Other distributions: 
\code{\link{dist_arcsine}()},
\code{\link{dist_exponential}()},
\code{\link{dist_frechet}()},
\code{\link{dist_gamma}()},
\code{\link{dist_lomax}()},
\code{\link{dist_soft_killing_constant_barrier}()},
\code{\link{dist_uniform}()},
\code{\link{dist_weibull}()}
}
\concept{distributions}
