% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iftp_distributions.R
\name{dist_lomax}
\alias{dist_lomax}
\title{Create a Lomax (Pareto Type II) distribution specification}
\usage{
dist_lomax(shape = 1, scale = 1)
}
\arguments{
\item{shape}{Shape parameter (alpha). Defaults to 1.}

\item{scale}{Scale parameter (sigma). Defaults to 1.}
}
\value{
An \code{iftp_dist} object with \code{cdf} and \code{quantile} functions.
}
\description{
CDF: \code{F(x) = 1 - (1 + x/scale)^(-shape)} for \code{x > 0}.
}
\examples{
d <- dist_lomax(shape = 2, scale = 1)
d$cdf(1)

}
\seealso{
Other distributions: 
\code{\link{dist_arcsine}()},
\code{\link{dist_exponential}()},
\code{\link{dist_frechet}()},
\code{\link{dist_gamma}()},
\code{\link{dist_log_logistic}()},
\code{\link{dist_soft_killing_constant_barrier}()},
\code{\link{dist_uniform}()},
\code{\link{dist_weibull}()}
}
\concept{distributions}
