% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iftp_distributions.R
\name{dist_uniform}
\alias{dist_uniform}
\title{Create a uniform distribution specification}
\usage{
dist_uniform(min = 0, max = 1)
}
\arguments{
\item{min}{Lower bound. Defaults to 0.}

\item{max}{Upper bound. Defaults to 1.}
}
\value{
An \code{iftp_dist} object with \code{cdf} and \code{quantile} functions.
}
\description{
Create a uniform distribution specification
}
\examples{
d <- dist_uniform(min = 0, max = 2)
d$cdf(1)

}
\seealso{
Other distributions: 
\code{\link{dist_arcsine}()},
\code{\link{dist_exponential}()},
\code{\link{dist_frechet}()},
\code{\link{dist_gamma}()},
\code{\link{dist_log_logistic}()},
\code{\link{dist_lomax}()},
\code{\link{dist_soft_killing_constant_barrier}()},
\code{\link{dist_weibull}()}
}
\concept{distributions}
