% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_checks.R
\name{is_iftp_dist}
\alias{is_iftp_dist}
\title{Test if an object is an IFTP distribution}
\usage{
is_iftp_dist(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"iftp_dist"}, \code{FALSE} otherwise.
}
\description{
Test if an object is an IFTP distribution
}
\examples{
is_iftp_dist(dist_exponential(1))
is_iftp_dist(pexp)
}
\seealso{
Other type-checks: 
\code{\link{is_iftp_process}()},
\code{\link{is_iftp_result}()},
\code{\link{is_ps_result}()}
}
\concept{type-checks}
