% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_checks.R
\name{is_iftp_process}
\alias{is_iftp_process}
\title{Test if an object is an IFTP process}
\usage{
is_iftp_process(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"iftp_process"}, \code{FALSE} otherwise.
}
\description{
Test if an object is an IFTP process
}
\examples{
is_iftp_process(process_bm())
is_iftp_process(rnorm)
}
\seealso{
Other type-checks: 
\code{\link{is_iftp_dist}()},
\code{\link{is_iftp_result}()},
\code{\link{is_ps_result}()}
}
\concept{type-checks}
