% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_checks.R
\name{is_iftp_result}
\alias{is_iftp_result}
\title{Test if an object is an IFTP result}
\usage{
is_iftp_result(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"iftp_result"}, \code{FALSE} otherwise.
}
\description{
Test if an object is an IFTP result
}
\examples{
is_iftp_result(iftp_anulova(dist_exponential(1),
  time_horizon = 1,
  time_steps = 5, n_particles = 50, seed = 1
))
is_iftp_result("not a result")
}
\seealso{
Other type-checks: 
\code{\link{is_iftp_dist}()},
\code{\link{is_iftp_process}()},
\code{\link{is_ps_result}()}
}
\concept{type-checks}
