% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_checks.R
\name{is_ps_result}
\alias{is_ps_result}
\title{Test if an object is a particle system result}
\usage{
is_ps_result(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"ps_result"}, \code{FALSE} otherwise.
}
\description{
Test if an object is a particle system result
}
\examples{
is_ps_result(nbmp_simulate(n_particles = 10, time_horizon = 0.5, seed = 1))
is_ps_result("not a result")
}
\seealso{
Other type-checks: 
\code{\link{is_iftp_dist}()},
\code{\link{is_iftp_process}()},
\code{\link{is_iftp_result}()}
}
\concept{type-checks}
