% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_result.R
\name{print.ps_result}
\alias{print.ps_result}
\title{Print a particle system result}
\usage{
\method{print}{ps_result}(x, ...)
}
\arguments{
\item{x}{A \code{ps_result} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a particle system result
}
