% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{shiny_generate_code}
\alias{shiny_generate_code}
\title{Generate reproducible R code snippet for a simulation}
\usage{
shiny_generate_code(params)
}
\arguments{
\item{params}{Named list with \code{sim_type} and code-generation strings.}
}
\value{
A character string of executable R code.
}
\description{
Generate reproducible R code snippet for a simulation
}
\keyword{internal}
